requirements.txt
pandas
numpy
scikit-learn
openai
streamlit
requests
google-generativeai
